les flags avancent...

+++ gettype.c
@@ -6,7 +6,7 @@
 /*   By: tfontain <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/07 21:29:39 by tfontain          #+#    #+#             */
-/*   Updated: 2017/03/18 13:01:29 by tfontain         ###   ########.fr       */
+/*   Updated: 2017/03/18 13:14:57 by tfontain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,14 +97,14 @@ ssize_t				ft_printdata(const char *t, uintmax_t data, int fd,
 		{
 			data = (uintmax_t)(data == 0 ? ft_memdup("(null)", 7)
 					: ft_memdup((void*)data, ft_strlen((char*)data)));
-			if (precision < ft_strlen((char*)data))
+			if (precision < (int)ft_strlen((char*)data))
 				((char*)data)[precision] = 0;
 		}
 		else
 		{
 			data = (uintmax_t)(data == 0 ? ft_memdup("L(null)", 28)
 					: ft_memdup((void*)data, ft_strwlen((wchar_t*)data)));
-			if (precision <= ft_strwlen((wchar_t*)data))
+			if (precision <= (int)ft_strwlen((wchar_t*)data))
 				((wchar_t*)data)[precision] = 0;
 		}
 		ret = ft_type(t)(data, fd);
